# openAPI-def.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "doughleap",
    "version": "1.0.0"
  },
  "paths": {
...
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get orders",
        "operationId": "getOrders",
        "parameters": [
          {
            "name": "orderIDs",
            "in": "query",
            "description": "Comma-separated list of order IDs",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "clientID",
            "in": "query",
            "description": "ID of the client",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "name of the order",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scheduledDeliveryTimeRange",
            "in": "query",
            "description": "Range of scheduled delivery dates to filter by (YYYY-MM-DD format)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "date-time"
              },
              "style": "form",
              "explode": true
            }
          },
          {
            "name": "createdTimeRange",
            "in": "query",
            "description": "Range of create dates to filter by (YYYY-MM-DD format)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "date-time"
              },
              "style": "form",
              "explode": true
            }
          },
          {
            "name": "fulfilledTimeRange",
            "in": "query",
            "description": "Range of fulfillment dates to filter by (YYYY-MM-DD format)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "date-time"
              },
              "style": "form",
              "explode": true
            }
          },
          {
            "name": "fulfillment",
            "in": "query",
            "description": "Order fulfillment type",
            "schema": {
              "type": "string",
              "enum": [
                "pickup",
                "delivery"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Order status",
            "schema": {
              "type": "string",
              "enum": [
                "created",
                "inProgress",
                "readyForDelivery",
                "delivered",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "orderID": {
                        "type": "integer"
                      },
                      "clientID": {
                        "type": "integer"
                      },
                      "scheduledDeliveryTime": {
                        "type": "string",
                        "format": "date"
                      },
                      "createdTime": {
                        "type": "string",
                        "format": "date"
                      },
                      "fulfilledTime": {
                        "type": "string",
                        "format": "date"
                      },
                      "fulfillment": {
                        "type": "string",
                        "enum": [
                          "pickup",
                          "delivery"
                        ]
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "created",
                          "inProgress",
                          "readyForDelivery",
                          "delivered",
                          "cancelled"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create an order",
        "operationId": "createOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "clientID",
                  "scheduledDeliveryTime",
                  "fulfillment",
                  "name"
                ],
                "properties": {
                  "clientID": {
                    "type": "integer"
                  },
                  "scheduledDeliveryTime": {
                    "type": "string",
                    "format": "date"
                  },
                  "fulfillment": {
                    "type": "string",
                    "enum": [
                      "pickup",
                      "delivery"
                    ]
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderID": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/orders/{orderID}": {
      "patch": {
        "tags": [
          "Orders"
        ],
        "summary": "Update an order",
        "operationId": "updateOrder",
        "parameters": [
          {
            "name": "orderID",
            "in": "path",
            "description": "ID of the order to update",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scheduledDeliveryTime": {
                    "type": "string",
                    "format": "date",
                    "description": "Updated delivery date (YYYY-MM-DD format)"
                  },
                  "description": {
                    "type": "string",
                    "description": "Updated description"
                  },
                  "name": {
                    "type": "string",
                    "description": "Updated name"
                  },
                  "fulfillment": {
                    "type": "string",
                    "enum": [
                      "pickup",
                      "delivery"
                    ],
                    "description": "Updated fulfillment type"
                  },
                  "fulfilledTime": {
                    "type": "string",
                    "format": "date",
                    "description": "Updated fulfilled date (YYYY-MM-DD format)"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "created",
                      "inProgress",
                      "readyForDelivery",
                      "delivered",
                      "cancelled"
                    ],
                    "description": "Updated status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderID": {
                      "type": "integer"
                    },
                    "clientID": {
                      "type": "integer"
                    },
                    "scheduledDeliveryTime": {
                      "type": "string",
                      "format": "date"
                    },
                    "createdTime": {
                      "type": "string",
                      "format": "date"
                    },
                    "fulfilledTime": {
                      "type": "string",
                      "format": "date"
                    },
                    "fulfillment": {
                      "type": "string",
                      "enum": [
                        "pickup",
                        "delivery"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "created",
                        "inProgress",
                        "readyForDelivery",
                        "delivered",
                        "cancelled"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "Delete an order",
        "operationId": "deleteOrder",
        "parameters": [
          {
            "name": "orderID",
            "in": "path",
            "description": "ID of the order to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Order deleted successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/tags": {
      "get": {
        "tags": [
          "Tags"
        ],
        "operationId": "getTags",
        "parameters": [
          {
            "name": "tagIDs",
            "in": "query",
            "description": "Comma-separated list of tag IDs",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "The name of the tag",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "tagID": {
                        "type": "integer",
                        "description": "The ID of the tag."
                      },
                      "name": {
                        "type": "string",
                        "description": "The name of the tag."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tags"
        ],
        "operationId": "createTag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The name of the tag."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tagID": {
                      "type": "integer",
                      "description": "The ID of the tag."
                    },
                    "name": {
                      "type": "string",
                      "description": "The name of the tag."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tags/{tagID}": {
      "patch": {
        "tags": [
          "Tags"
        ],
        "operationId": "updateTag",
        "parameters": [
          {
            "name": "tagID",
            "in": "path",
            "description": "The ID of the tag.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "The updated name of the tag.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "description": "The name of the tag."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tagID": {
                      "type": "integer",
                      "description": "The ID of the tag."
                    },
                    "name": {
                      "type": "string",
                      "description": "The name of the tag."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tags"
        ],
        "operationId": "deleteTag",
        "parameters": [
          {
            "name": "tagID",
            "in": "path",
            "description": "The ID of the tag.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
...
}